requirements.txt
langchain
ipykernel
python-dotenv
huggingface_hub
transformers
langchain_huggingface
openai
google.generativeai